third_party/WebKit/LayoutTests/bluetooth/getPrimaryService/delayed-discovery-service-not-found.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('DelayedServicesDiscoveryAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['battery_service']}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      return assert_promise_rejects_with_message(
        gattServer.getPrimaryService('battery_service'),
        new DOMException('Service not found in device.', 'NotFoundError'));
    });
}, 'Request for absent service. Must reject with NotFoundError even when the ' +
   'services are not immediately discovered');
</script>
